requirements_decoder.txt
pytorch-lightning
scikit-learn
einops
tiktoken
numba
pandas
h5py
transformers
tensorboardX
wandb
